requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
firebase-admin>=6.0.0
python-dotenv>=1.0.0
gunicorn>=20.1.0
whitenoise>=6.0.0